+++ src/atoms/list-atom/listBuilder.ts
@@ -1,36 +1,33 @@
 import { FieldAtom, FormFieldValues, FormFields } from "form-atoms";
 
-type EmptyValues<Fields extends FormFields> = {
-  [Key in keyof Fields]: Fields[Key] extends FieldAtom<any>
-    ? undefined
-    : Fields[Key] extends FormFields
-      ? EmptyValues<Fields[Key]>
-      : Fields[Key] extends Array<infer Item>
-        ? Item extends FieldAtom<any>
-          ? undefined[]
-          : Item extends FormFields
-            ? EmptyValues<Item>[]
-            : never
-        : never;
-};
+import { FormFieldSubmitValues } from "../../components/form";
+import { ZodField, ZodFieldSubmitValue } from "../../fields";
 
-type ListFieldItems = FieldAtom<any> | FormFields;
+export type ListAtomItems = FieldAtom<any> | FormFields;
 
-type ListFieldValues<T> = T extends FieldAtom<infer Value>
-  ? Value | undefined // also empty
+export type ListAtomValue<T extends ListAtomItems> = T extends FieldAtom<
+  infer Value
+>
+  ? Value
   : T extends FormFields
-    ? FormFieldValues<T> | EmptyValues<T>
+    ? FormFieldValues<T>
+    : never;
+
+export type ListAtomSubmitValue<T extends ListAtomItems> = T extends ZodField
+  ? ZodFieldSubmitValue<T>
+  : T extends FormFields
+    ? FormFieldSubmitValues<T>
     : never;
 
 // actual type must be one of overloads, as this one is ignored
-export function listFieldBuilder<
-  Fields extends ListFieldItems,
+export function listBuilder<
+  Fields extends ListAtomItems,
   /**
    * HACK
    * Having the Values computed in generic argument,
    * fixes the return types **when the argument to builder is used**.
    */
-  Values extends ListFieldValues<Fields>,
+  Values extends ListAtomValue<Fields>,
 >(builder: (value: Values) => Fields) {
   let emptyValue: undefined | Values = undefined;
   try {
@@ -48,7 +45,7 @@ export function listFieldBuilder<
    * HACK2:
    * the data is not simply Values, as that would produce the any[] on the returned function.
    */
-  function buildFields(data: ListFieldValues<Fields>[]): Fields[];
+  function buildFields(data: ListAtomValue<Fields>[]): Fields[];
   function buildFields(data?: Values[]) {
     if (data) {
       return data.map(builder);