+++ src/fields/digit-field/digitField.test-d.ts
@@ -2,7 +2,7 @@ import { formAtom } from "form-atoms";
 import { expectTypeOf, test } from "vitest";
 
 import { digitField } from "./digitField";
-import { FormSubmitValues } from "../zod-field/zodField";
+import { FormSubmitValues } from "../../components/form";
 
 test("required digitField has '0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9' submit value", () => {
   const form = formAtom({