+++ src/fields/zod-field/zodField.test.ts
@@ -1,17 +1,30 @@
 import { act, renderHook } from "@testing-library/react";
 import { useFieldActions } from "form-atoms";
 import { describe, expect, test } from "vitest";
+import { z } from "zod";
 
+import { zodField } from "./zodField";
 import { useRequiredProps } from "../../hooks";
-import { stringField } from "../string-field";
 
 describe("zodField", () => {
-  describe(".optional(readRequired)", () => {
-    test("it makes the required field state depend on other atoms", () => {
-      const customerType = stringField({ value: "personal" });
+  describe("optional()", () => {
+    test("returns the same field when calling optional", () => {
+      const field = zodField({
+        value: undefined,
+        schema: z.number(),
+      }).optional();
 
-      const vatNumber = stringField().optional(
-        (get) => get(get(customerType).value) === "business",
+      const fieldRef = field.optional().optional();
+
+      expect(fieldRef).toEqual(field);
+    });
+  });
+
+  describe("callback argument", () => {
+    test("enables the optional field to be conditionally required", () => {
+      const accountType = zodField({ value: "personal", schema: z.string() });
+      const vatNumber = zodField({ value: "", schema: z.string() }).optional(
+        (get) => get(get(accountType).value) === "business",
       );
 
       const requiredProps = renderHook(() =>
@@ -20,7 +33,7 @@ describe("zodField", () => {
 
       expect(requiredProps.result.current.required).toBe(false);
 
-      const customerActions = renderHook(() => useFieldActions(customerType));
+      const customerActions = renderHook(() => useFieldActions(accountType));
 
       act(() => customerActions.result.current.setValue("business"));