+++ src/fields/zod-field/zodField.ts
@@ -1,67 +1,26 @@
-import { FieldAtom, FieldAtomConfig, FormAtom, fieldAtom } from "form-atoms";
-import { zodValidate } from "form-atoms/zod";
-import { Atom, Getter, WritableAtom, atom } from "jotai";
-import { RESET } from "jotai/utils";
+import { FieldAtom, FieldAtomConfig, fieldAtom } from "form-atoms";
+import { Atom } from "jotai";
 import { ZodAny, ZodUndefined, z } from "zod";
 
-type ValidationConfig<Schema extends z.Schema, OptSchema extends z.Schema> = {
-  schema: Schema | ((get: Getter) => Schema);
-  optionalSchema?: OptSchema | ((get: Getter) => OptSchema);
-};
-
-type Flatten<T> = Identity<{
-  [K in keyof T]: T[K];
-}>;
-type Identity<T> = T;
-
-export type FormSubmitValues<Form extends FormAtom<any>> =
-  Form extends FormAtom<infer Fields> ? FormFieldSubmitValues<Fields> : never;
-
-type FormFieldSubmitValues<Fields extends FormFields> = Flatten<{
-  [Key in keyof Fields]: Fields[Key] extends ZodField<any>
-    ? ZodFieldSubmitValue<Fields[Key]>
-    : Fields[Key] extends FieldAtom<infer Value>
-      ? Value
-      : Fields[Key] extends FormFields
-        ? FormFieldSubmitValues<Fields[Key]>
-        : Fields[Key] extends Array<infer Item>
-          ? Item extends ZodField<any>
-            ? ZodFieldSubmitValue<Fields[Key]>[]
-            : Item extends FieldAtom<infer Value>
-              ? Value[]
-              : Item extends FormFields
-                ? FormFieldSubmitValues<Item>[]
-                : never
-          : never;
-}>;
-
-type FormFields = {
-  [key: string | number]:
-    | FieldAtom<any>
-    | ZodField<any>
-    | FormFields
-    | FormFields[]
-    | FieldAtom<any>[]
-    | ZodField<any>[];
-};
+import { ExtendFieldAtom, extendFieldAtom } from "../../atoms/extendFieldAtom";
+import {
+  ReadRequired,
+  ValidateConfig,
+  WritableRequiredAtom,
+  schemaValidate,
+} from "../../atoms/schemaValidate";
 
 export type ZodFieldConfig<
   Schema extends z.Schema,
   OptSchema extends z.Schema = ZodUndefined,
 > = FieldAtomConfig<Schema["_output"] | OptSchema["_output"]> &
-  ValidationConfig<Schema, OptSchema>;
+  ValidateConfig<Schema, OptSchema>;
 
 export type ZodFieldValue<Field> = Field extends FieldAtom<infer Value>
   ? Value
   : never;
 
-type WritableRequiredAtom = WritableAtom<
-  boolean,
-  [boolean | typeof RESET | ((prev: boolean) => boolean)],
-  void
->;
-
-type ZodFieldSubmitValue<Field> = Field extends ZodField<
+export type ZodFieldSubmitValue<Field> = Field extends ZodField<
   infer Schema,
   infer OptSchema,
   infer Required
@@ -86,12 +45,6 @@ type RequiredZodField<
   OptSchema extends z.Schema = ZodUndefined,
 > = ZodField<Schema, OptSchema>;
 
-type ExtendFieldAtom<Value, State> = FieldAtom<Value> extends Atom<
-  infer DefaultState
->
-  ? Atom<DefaultState & State>
-  : never;
-
 export type ZodField<
   Schema extends z.Schema = ZodAny,
   OptSchema extends z.Schema = ZodUndefined,
@@ -101,105 +54,44 @@ export type ZodField<
   { required: RequiredAtom }
 > & {
   optional: (
-    read?: Atom<boolean>["read"],
+    readRequired?: ReadRequired,
   ) => OptionalZodField<Schema, OptSchema>;
 };
-
-/**
- * Read-only atom for default zodFields which all are required.
- */
-const defaultRequiredAtom = atom(true as const);
-defaultRequiredAtom.debugLabel = "zodField/defaultRequired";
 
 export function zodField<
   Schema extends z.Schema,
   OptSchema extends z.Schema = ZodUndefined,
->({
-  schema,
-  optionalSchema,
-  ...config
-}: ZodFieldConfig<Schema, OptSchema>): RequiredZodField<Schema, OptSchema> {
-  const baseFieldAtom = fieldAtom({
-    validate: zodValidate(
-      (get) => {
-        const schemaObj = typeof schema === "function" ? schema(get) : schema;
-        return schemaObj;
-      },
-      {
-        on: "blur",
-        when: "dirty",
-      },
-    ).or({ on: "change", when: "touched" }),
-    ...config,
+>({ schema, optionalSchema, ...config }: ZodFieldConfig<Schema, OptSchema>) {
+  const { validate, requiredAtom, makeOptional } = schemaValidate({
+    schema,
+    optionalSchema,
   });
 
-  const zodField = atom((get) => {
-    const baseField = get(baseFieldAtom);
+  const zodFieldAtom = extendFieldAtom(fieldAtom({ ...config, validate }), {
+    required: requiredAtom,
+  }) as unknown as RequiredZodField<Schema, OptSchema>;
 
-    const fieldAtoms = {
-      required: defaultRequiredAtom,
-    };
+  zodFieldAtom.optional = (readRequired: ReadRequired = () => false) => {
+    const { validate, requiredAtom } = makeOptional(readRequired);
 
-    return { ...baseField, ...fieldAtoms };
-  }) as unknown as ZodField<Schema, OptSchema>;
+    const optionalZodFieldAtom = extendFieldAtom(
+      fieldAtom({ ...config, validate }),
+      { required: requiredAtom },
+    ) as OptionalZodField<Schema, OptSchema>;
 
-  const makeOptional = (readRequired: Atom<boolean>["read"] = () => false) => {
-    const requiredAtom = atom(readRequired);
+    optionalZodFieldAtom.optional = () => optionalZodFieldAtom;
 
-    const baseFieldAtom = fieldAtom({
-      validate: zodValidate(
-        (get) => {
-          const schemaObj = typeof schema === "function" ? schema(get) : schema;
-
-          const optionalSchemaObj =
-            typeof optionalSchema === "function"
-              ? optionalSchema(get)
-              : optionalSchema;
-
-          const optSchema = optionalSchemaObj ?? schemaObj.optional();
-
-          const isRequired = get(requiredAtom);
-
-          return isRequired ? schemaObj : optSchema;
-        },
-        {
-          on: "blur",
-          when: "dirty",
-        },
-      ).or({ on: "change", when: "touched" }),
-      ...config,
-    });
-
-    const optionalZodField = atom((get) => {
-      const baseField = get(baseFieldAtom);
-
-      const fieldAtoms = {
-        required: requiredAtom,
-      };
-
-      if (
-        typeof process !== "undefined" &&
-        process.env.NODE_ENV !== "production"
-      ) {
-        Object.entries(fieldAtoms).map(([atomName, atom]) => {
-          atom.debugLabel = `field/${atomName}/${config.name ?? zodField}`;
-        });
-      }
-
-      return {
-        ...baseField,
-        ...fieldAtoms,
-      };
-    }) as OptionalZodField<Schema, OptSchema>;
-
-    optionalZodField.optional = () => optionalZodField;
-    optionalZodField.debugLabel = `optionalZodField/${config.name ?? zodField}`;
-
-    return optionalZodField;
+    return optionalZodFieldAtom;
   };
 
-  zodField.optional = makeOptional;
-  zodField.debugLabel = `zodField/${config.name ?? zodField}`;
-
-  return zodField;
+  return zodFieldAtom;
 }
+
+// if (
+//   typeof process !== "undefined" &&
+//   process.env.NODE_ENV !== "production"
+// ) {
+//   Object.entries(fieldAtoms).map(([atomName, atom]) => {
+//     atom.debugLabel = `field/${atomName}/${config.name ?? zodField}`;
+//   });
+// }