+++ src/hooks/index.ts
@@ -5,7 +5,7 @@ export * from "./use-date-field-props";
 export * from "./use-field-error";
 export * from "./use-field-props";
 export * from "./use-files-field-props";
-export * from "./use-list-field-actions";
+export * from "./use-list-field";
 export * from "./use-multiselect-field-props";
 export * from "./use-number-field-props";
 export * from "./use-options";